Fix Visualization of multiple lines
Important commit, useful for future reference.  The problem showed when
you added a new project or metric to be visualized.  The data would be
fetched and sent to the Vega visualizer.  However, the data was already
parsed by the Vega visualization and so a second round of parsing only
did bad things.  Furthermore, it seems that the vega parsing caused
knockout to think some dependency was being updated and it re-calculated
the datasets and re-drew the graph unnecessarily.

Fixes to both of these sub-issues are in this patch.

Change-Id: I8c7816b6321cefddc70ba07f782996464bbb7159