Fix metric and project selection
When adding a metric, make it the selected metric
When adding a project that has no language choice (e.g., Commons),
we should not ask the user to click on "All languages"
but add "All languages" automatically.
Fix typeahead binding to only add a single typeahead instance (race
conditions were causing multiple instances to be added before this
change).

Change-Id: I1acce65c6a4fdb61c0a314f0778c7b0f3920e6df