Add Separated Values (CSV, TVS) data converter
A new converter also means the visualizers need to know which converter
to use. A new factory was added that abstracts that decision. Since so
far we have a different data transform per api we have retrofitted
the data apis with the information about which
converter/transformer to use.

Stubs have been added to be able to test UI
reponses and display for legacy pageview data.
Stubs can be removed as vetted data becomes available.

Change-Id: I85e89f5eb7dabb2785c2e1e2c0068fc3d938a7a1