Add annotations to graphs
This changeset uses annotations API to
get metric annotation json specifications from
mediawiki pages, and displays a list of annotations
at the bottom of each graph (if annotations available).

Each annotation consists of a box with the date range
in which the event took place, and a popup note with
the specified text.

Bug: T78151
Change-Id: I0b52abdce474255609982441d68448d6179d1a35