Prevent crashing ui when metric data is missing
Whenever any of the files that contain metric data failed to load
(for any reason) the ui was crashing and the other metrics were
not shown, even if their data had been retrieved normally.

This fix adresses this problem at API level. So, when an error
occurs fetching the data, the error is logged and the api returns
an empty list.

Bug: T85233
Change-Id: I4323e87922de83bc261ce34a5d067cc04ad3ef52