[Review but DO NOT MERGE] Begin funnel layout
This is the start of a series of patches that will implement a new
layout for the visual editor funnel analysis, in dashiki.  It's
exploratory only, and a proper refactor of this and the wikimetrics
layout into separate repositories should be done.

This patch can be reviewed, but should not be merged until the
appropriate refactoring is done.

Change-Id: Id732758508048e5b1d3126534ad145cb21bfa4f4