Make the gulp build layout specific
Instead of building /src with a hard-coded configuration, after this
change the user can run this command:

    gulp --layout compare --config SomeWikiArticle

What this will do is build the layout found in /src/layouts/compare, and
use meta.wikimedia.org/wiki/Config:SomeWikiArticle as configuration for
this layout.  It will put the output in /dist/compare-SomeWikiArticle
which is now ignored from the repository.  This directory is ready to be
served by an appropriately configured Apache site.  Since there is no
more hard-coded configuration, this opens up dashiki for much wider use
and customization.

Status of change:
* /src and /dist modes of both layouts are working.
* configuration for both layouts is external
* require js optimizer is working for both layouts
* testing is only partially complete for new components

In my opinion, it's ready to merge and we can clean up tests as we
further modularize it and break down the components.

Bug: T96337
Change-Id: I704e86c7170c9de3501d06c0c9d4072860fed4de