Sort the legend by value and align it
mforns had this idea of sorting the legend in descending order and I
ignored it because it would've required a big hack.  It turns out that
styling the legend labels requires the big hack anyway, so I did it and
I love it, it makes the graph make so much more sense.

Bug: T131935
Change-Id: I93a749f4d354d1e2ca1ba7c95ef1c21efb19cabd