Fix unique devices bugs. Update to knockout 3.4
After trying to deploy to production we spotted a couple bugs.
We rolled back to the latest stable commit.

This patch fixes the found bugs:

1) The format of the date was wrong for some granularities.
For example, daily granularity would send date strings with
hour values. Also, monthly granularity would send daily values.
This was accepted in some cases by AQS but in others returned
404. To reproduce, just select Monthly Unique Devices and
look at the network tab in the browser's developer tools.
To solve this, the config now contains date formats per
granularity.

2) In some cases, breakdown-toggle would try to iterate the
metric breakdown before the metric or its breakdown was
initialized, and this would generate an exception (race condition).
To reproduce, try to change between breakdownable metrics a couple
times. To solve this, I added an if statement that checks that
the metric and the breakdown are not undefined.

3) Sometimes the breakdowns in the left menu would be duplicated.
To reproduce this, add a metric with breakdown, like Pageviews.
Then add a metric without them, like Rolling Active Editor.
Then go back to the breakdown metric and activate the breakdown.
You should see duplicated (or triplicated, etc) options.
To solve this several changes were needed among them the usage of
knockout deferred updates to avoid triggering a cascade of
computed observables changes when one obserbable changes. Using deferred
updates we avoid intermediate computations that are not necessary.

Bug: T122533
Change-Id: Ic0dee6ca7f5514fba0f3fe9638ae2db9d4a47a89