MonthlyUniqueDevices metric should be bookmarkeable
The Monthly unique devices metric was not bookmarkeable
as it shared the metric "name" with UniqueDevices. It only had
a distinct display name. The actual metric name (and display name)
need to be different so both metrics can be bookmarked
independently. The fact that both metrics come from
the same endpoint is handled by the config

Bug: T122533
Change-Id: I76a2fda8af61710ef78f4cca9f70e27da45d4588