Migrate from bower to npm instead of yarn
We originally thought migrating to yarn would be a good idea.  It turns
out it fails to install some bower packages and doesn't work well with
git repository dependencies yet.  So we'll include the yarn.lock file
for now and look again in a few months when yarn 1.0 is released.  Until
then we have moved to npm from bower, simplifying and upgrading the
packages.

The google fonts kept breaking the build in offline mode and otherwise,
so we also spent some time and replaced them with locally downloaded
fonts.  This means certain files of the semantic build had to be checked
into source control and the build had to be changed a bit.  All is well.

Bug: T147884
Change-Id: Icd28e93117370d3cbb151cda22a15c69f83e58d8