Fixing tests on master
Issue 1:
Turns out that scoping in jasmine is not what you think it is:
https://github.com/jasmine/jasmine/issues/941

Fixed variables so they are locally scoped

Issue 2:
The introduction on phantomjs brough some issues
when it comes to mocking/sinon and support. Phantom does not support
new APIs as well as other browsers
so we might want to rethink the choice of using it.
Rolled back sinon to the prior minor release that works better

https://github.com/sinonjs/sinon/issues/228

Bug: T152631

Change-Id: I509d0596cdfa7e3b8abd6e545799a58b6c9cc415