Remove decomissioned dashboards
Note: the reportupdater jobs generating data for these dashboards will
be shut down as well, and the dashboards will be taken offline effective
immediately.  We will leave the data in place for now, and leave a note
pointing to it from the old dashboard urls, in case people want to track
it down for historical purposes.

Bug: T199340
Change-Id: Ice7c0328dc096947d31b6e831ae3a6b7f09107bc