Fix pageviews.js reference
Weirdest bug ever.  Pageviews.js has a weird call and console.log at the
end of its definition.  It's not present in the minified file, so the
fix is to simply use that.  But if you look in the source, the debug
code is not there, it only shows up when you npm install and look in
node_modules/pageviews/pageviews.js

Also, update semantic in its json config.

Change-Id: I3e1378e6db0a3deb4f7b56fb343d18904a8df840