Add grouped option to datasets api
If grouped is set, instead of fetching each requested project
separately, we fetch a file called grouped.{format} and pre-parse it.
We extract from it all selected wikis and ignore the rest.  This adds
overhead on the client but makes it much easier to output from Hive
where the wiki is usually part of the select statement.

Some conventions are added here, and will be documented along with new
documentation for schemas in a following change.  Conventions are:

The file containing all wikis must be called {submetric}.{format}
The grouped file must have three columns: wiki, date, total

Note: I added a couple of tests but testing is completely broken in
dashiki so I'm leaving that fix for another time.

Bug: T236941
Change-Id: I323f17589d5c9b636908cfab950e6341962e6324