Fix gulp build
This goes with the general dependency upgrade.  Set all build
dependencies to latest possible version and fixed any bugs.  Gulp
tasks required some updating and as a bonus the linter is better so we
cleaned up some js too.

Testing is still broken, but I'm tempted to leave it like that for now.
It's a pain to maintain karma, we might want to upgrade to jest, rollup,
etc.

Change-Id: I38a457ca2fecb16a731578907322e11e4a2cd718