Fix gulp build properly this time
The js task was not working because the requirejs-bundler is a very old
plugin that is not using the newest Vinyl file format.  I made a little
adapter gulp plugin that sits on the stream and just converts the file
objects and it seems to work.  I also moved around all the task code
because the other tasks take negligible time and it was hard to debug
the js task with the previous ordering.

Change-Id: I9170e5187fc116fdcd5f4f4515a508cf408f98e6